Refactor: Move types and utilities to shared directory; update imports
