+++ app/auth/update-password/page.tsx
@@ -16,7 +16,7 @@ import {
 import { Alert, AlertDescription } from "@/components/ui/alert";
 import { CheckCircle, AlertCircle, Loader2, Lock } from "lucide-react";
 import { useRouter } from "next/navigation";
-import { validatePassword } from "@/lib/validation";
+import { validatePassword } from "@/lib/shared/validation";
 import { updatePassword } from "@/app/auth/update-password/actions";
 
 export default function UpdatePasswordPage() {
