+++ components/private/chat/ChatMessages.tsx
@@ -3,7 +3,7 @@ import { motion, AnimatePresence } from "framer-motion";
 import MarkdownTypewriter from "@/components/ui/MarkdownTypewriter";
 import MarkdownRenderer from "@/components/ui/MarkdownRenderer";
 import { CompareResponses } from "@/components/private/modes/Compare/CompareResponses";
-import type { Message } from "@/lib/types";
+import type { Message } from "@/lib/shared/types";
 
 interface ChatMessagesProps {
   messages: Message[];
