+++ lib/auth/auth-errors.ts
