+++ lib/shared/utils.ts
