+++ lib/shared/validation.ts
